+++ CHANGES.rst
@@ -1,6 +1,8 @@
 0.7 (unreleased)
 ----------------
 
+- ``plot_fit`` and ``plot_data`` have new options to control the marker size of
+  spectral flux points and the shape of upper-limit arrows.
 
 0.6.1 (2015-10-29)
 ------------------